requirements.txt
flask==3.0.3
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
